fix tests; fix elements: line, transformer ends, tap changers

+++ cgmes/cgmes-update/src/main/java/com/powsybl/cgmes/update/elements14/ShuntCompensatorToShuntCompensator.java
@@ -36,7 +36,7 @@ public class ShuntCompensatorToShuntCompensator implements ConversionMapper {
         map.put("bPerSection", new CgmesPredicateDetails(
             "cim:ShuntCompensator.bPerSection", "_EQ", false, String.valueOf(bPerSection)));
 
-        double maximumSectionCount = newShuntCompensator.getMaximumSectionCount();
+        int maximumSectionCount = newShuntCompensator.getMaximumSectionCount();
         map.put("maximumSectionCount", new CgmesPredicateDetails(
             "cim:ShuntCompensator.maximumSections", "_EQ", false, String.valueOf(maximumSectionCount)));
